Week_8/requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
faiss-cpu>=1.8.0
sentence-transformers>=2.5.0
requests>=2.31.0
numpy>=1.26.0
scikit-learn>=1.4.0
